Use map instead of array for storing cached papers.
There were some bugs before because `paper_data` was accessed using an index instead of `filtered_paper_data` (or vice versa, I don't quite recall). There didn't seem to be a good way to get the proper indices into the original data given the filtered data, so I've just changed it so the original data is an object (map) with paper ids as keys. Now it's easy to edit the map given a paper, and I've changed the places where indices were passed around to just pass the paper instead.

The only potential downside I see is that we can't (at least, that I'm aware of) order the papers in the map by priority, so we have to sort the filtered papers each time. This shouldn't take long enough for it to be a concern, though.