Link page for writing notes to papers.
Now you can click a new button to pull up the dedicated notes page for a paper and it will pre-populate with the paper's abstract. Saving (which is done automatically every 30s or by clicking a button or ctrl+s) updates the paper in the database.

As the Notes page depends on knowing what paper you're editing, I removed it from the nav drawer.