Md: moved preprocessing / rendering to own file.
This makes it easier to have a consistent markdown rendering experience between different components (e.g. this fixes some differences between the rendering options used for the expansion panels and the notes page).

I've removed image cache clearing for now until I get a chance to test how much time is spent adding things to the cache / how much memory it uses. This also makes it possible to see the images in the expansion panels because there's no way to re-render there (a future update should force a re-render once images are loaded so that cache clearing on exiting the Notes page wouldn't make images unviewable in the expansion panels).