Use git (untested), separate SavedQuery from other entries.
SavedQuery has been made an entry so they can be saved/loaded the same way, but it needs to be treated specially in a couple of places. I've removed it from the dropdown of entry types when creating a SavedQuery and also from the auto-generated "All Entrys" SavedQuerys (so you don't get a SavedQuery to show you the SavedQuerys).

The git functions have implementations now, but they haven't been tested yet.