git: add files to index instead of "soft" committing.
Since we were just amending the "soft" commit throughout the day, it didn't really need to be a commit at all, just a list of things we want to commit once it's time (e.g. once it's a new day). So to save time and make things simpler, soft committing is replaced with adding files to the index. There's a new function which will commit only if the previous commit was yesterday or earlier. This also lets us remove all of the custom git reset stuff.