git: show slider on Notes page to view previous versions.
The UI / UX here needs work, but the functionality seems good. You can move back to any previous commit and it will show a non-editable textarea for the previous markdown and a render of it. Both are scrollable and look like the editable version except with the text gray to hint to the user that editing is disabled.